Added Loading page
